+++ src/Model/Reports/WorkloadReportWorker.php
@@ -10,9 +10,13 @@ class WorkloadReportWorker extends Worker
     /** @var ArrayCollection<int, float> */
     public ArrayCollection $loggedPercentage;
 
+    /** @var float */
+    public float $average;
+
     public function __construct()
     {
         parent::__construct();
+        $this->average = 0.0;
         $this->loggedPercentage = new ArrayCollection();
     }
 }